+++ Scene/Vector.cpp
@@ -11,19 +11,19 @@ Vector::Vector()
 }
 
 
-Vector::Vector(double _x, double _y, double _z)
+Vector::Vector(float _x, float _y, float _z)
 {
    set(_x, _y, _z, 0.0);
 }
 
 
-Vector::Vector(double _x, double _y, double _z, double _w)
+Vector::Vector(float _x, float _y, float _z, float _w)
 {
    set(_x, _y, _z, _w);
 }
 
 
-void Vector::set(double _x, double _y, double _z, double _w)
+void Vector::set(float _x, float _y, float _z, float _w)
 {
    x = _x;
    y = _y;
@@ -32,7 +32,7 @@ void Vector::set(double _x, double _y, double _z, double _w)
 }
 
 
-Vector Vector::times(double s) /* scalar times Vector */
+Vector Vector::times(float s) /* scalar times Vector */
 {
    return Vector(s*x, s*y, s*z, s*w);
 }
@@ -48,14 +48,14 @@ Vector Vector::crossProduct(Vector v)
    return Vector((y*v.z)-(z*v.y), (z*v.x)-(x*v.z), (x*v.y)-(y*v.x));
 }
 
-double Vector::dotProduct(Vector v)
+float Vector::dotProduct(Vector v)
 {
    return x*v.x + y*v.y + z*v.z;
 }
 
 Vector Vector::normalize()  /* return a vector with length 1 */
 {
-   double norm = sqrt( x*x + y*y + z*z );
+   float norm = sqrt( x*x + y*y + z*z );
    return Vector(x/norm, y/norm, z/norm);
 }
