+++ Scene/Vector.h
@@ -5,23 +5,23 @@ class Vector
 {
    private:
    public:
-      double x, y, z, w; /* a vector in homogenous coordinates */
+      float x, y, z, w; /* a vector in homogenous coordinates */
 
       Vector();
 
-      Vector(double x, double y, double z);
+      Vector(float x, float y, float z);
 
-      Vector(double x, double y, double z, double w);
+      Vector(float x, float y, float z, float w);
 
-      void set(double _x, double _y, double _z, double _w);
+      void set(float _x, float _y, float _z, float _w);
 
-      Vector times(double s); /* scalar times Vector */
+      Vector times(float s); /* scalar times Vector */
 
       Vector plus(Vector v);  /* Vector addition */
 
       Vector crossProduct(Vector v);
 
-      double dotProduct(Vector v);
+      float dotProduct(Vector v);
 
       Vector normalize();     /* return a vector with length 1 */
