+++ Scene/Vertex.h
@@ -7,38 +7,38 @@ class Vertex
 {
    private:
    public:
-      double x, y, z, w;  /* a vertex in homogenous coordinates */
-      double r, g, b, a;  /* color at this vertex (including transparency) */
-      double s, t;        /* texture coordinates at this vertex */
+      float x, y, z, w;  /* a vertex in homogenous coordinates */
+      float r, g, b, a;  /* color at this vertex (including transparency) */
+      float s, t;        /* texture coordinates at this vertex */
       Vector n;           /* normal vector at this vertex */
 
       Vertex();
 
-      Vertex(double x, double y, double z);
+      Vertex(float x, float y, float z);
 
-      Vertex(double x, double y, double z, double w);
+      Vertex(float x, float y, float z, float w);
 
       Vertex(float v[4]);
 
-      void set(double x, double y, double z);
+      void set(float x, float y, float z);
 
-      void set(double _x, double _y, double _z, double _w);
+      void set(float _x, float _y, float _z, float _w);
 
       /* sets the color and transparancy of this vertex to (r,g,b,a) */
-      void setColor(double _r, double _g, double _b, double _a);
+      void setColor(float _r, float _g, float _b, float _a);
 
-      void setColor(double _r, double _g, double _b);
+      void setColor(float _r, float _g, float _b);
 
       void setColor(int _r, int _g, int _b, int _a);
 
       void setColor(int _r, int _g, int _b);
 
       /* sets the texture coordinate of this vertex to (s, t) */
-      void setCoords(double _s, double _t);
+      void setCoords(float _s, float _t);
 
       void setNormal(Vector * _n);
 
-      Vertex times(double scalar);
+      Vertex times(float scalar);
 
       /* for debugging */
       void print(int fields);
